// wardn/frontend/lib/api/generated/model/secretStoreRead.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.1.0
 */
import type { SecretStoreReadAuthConfig } from './secretStoreReadAuthConfig';
import type { SecretStoreReadConfig } from './secretStoreReadConfig';

export interface SecretStoreRead {
  authConfig: SecretStoreReadAuthConfig;
  config: SecretStoreReadConfig;
  createdAt: string;
  createdById?: string | null;
  id: string;
  isActive: boolean;
  name: string;
  organizationId?: string | null;
  provider: string;
  updatedAt: string;
  workspaceId?: string | null;
}